Persia decision and event
Now the Soviet Union or United Kingdom can call each other to do a intervention in Persia and change their government. Persia, in other side, can choice if will surrender to demands or will resist.
If UK or SOV choice do not a intervention, and Germany conquest Satalingrad, Persia can use a diplomatic choice to enter in the war in German side and revindicate a old territories from Great Persia.